Modulize the printer function
Modulize the log printer function, to make it more consistent across scripts.
